Google writing exercise, documenting sample code
This Java method was part of a Google test for technical writers that I received in 2014 and 2017. Instructions at the time read as below:
This section contains a short code sample written in Java. Please do the following:

1. Assume the sample appears in a document for programmers. Describe the code in some detail. You can use text, code comments, or any other techniques you like.

2. Assume you have a chance to send comments to the person who wrote the code. Suggest ways to improve the code (for example, to reduce memory usage or enhance features).

Your writing should be ready for publication and should have a level of depth suitable for an experienced developer audience.