A complete Java console application
Modified and extended the Google `findNeedles()` method to include error handling and the ability to enter a target text file to search. Also added code to allow user to choose the number of 'needles' to look for (from 1 to 5). Improved the logic of the findNeedles() method, to avoid redundant processing.